chore(dashboard): unify on Preact, remove the SolidJS toolchain (#66)
The dashboard carried a second UI framework: a Solid "island"
(kpi-strip-island) that mounted a Solid subtree inside the Preact app
for fine-grained KPI reactivity, plus a dead headless-solid experiment
and a solid-poc preview with zero consumers. This split forced a
dual-transform build (vite-plugin-solid + preact), a second test config
(vitest.solid.config.ts), and per-component .solid mirrors.

Per the maintainer decision, unify on Preact:

- kpi-strip-island.ts now renders the Preact KpiStrip/KpiCell directly.
  This is the exact output the island's test-environment fallback
  already used, so the five call sites (governance, harness-health,
  feature-health, surface-readiness, connector-status) are unchanged.
- Remove the Solid mirror components and their tests: kpi-strip.solid,
  kpi-cell.solid, bar.solid, kpi-strip-island-solid, + *.solid.test.
- Remove the dead headless-solid/ experiment (3 hooks + tests, only
  consumer was the dead solid-poc) and the solid-poc preview.
- Drop solid-js, @solidjs/testing-library, vite-plugin-solid from
  package.json; remove the solid plugin + solid chunk from vite.config;
  delete vitest.solid.config.ts and its test-script leg.
- connector-status.test.ts no longer needs its Solid-avoidance mock —
  it exercises the real (now Preact) KpiStripIsland.

Trade-off: Solid diffed only changed cells; Preact re-renders the whole
strip per parent render. A KPI strip is a few cells, so the cost is
negligible, and the second framework's build/test complexity is gone.

Verification (local): tsc --noEmit clean; kpi-strip/kpi-cell/bar/
connector-status/governance/harness-health/feature-health/
surface-readiness tests pass (159 total). dune build runs in CI.
