--- dashboard/design-system/preview/solid-poc.html
@@ -1,63 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="utf-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <title>MASC — SolidJS PoC (RFC 0017)</title>
-  <link rel="stylesheet" href="../source_styles/tokens.generated.css" />
-  <link rel="stylesheet" href="_preview.css" />
-  <style>
-    body { padding: 20px; }
-    .solid-poc {
-      display: flex;
-      flex-direction: column;
-      gap: 16px;
-      max-width: 920px;
-      margin: 0 auto;
-      color: var(--color-fg-primary);
-    }
-    .solid-poc header h1 {
-      font-family: var(--font-mono);
-      font-size: 20px;
-      margin: 0 0 4px;
-    }
-    .solid-poc .muted {
-      color: var(--color-fg-muted);
-      font-size: var(--fs-12);
-    }
-    .controls { display: flex; gap: 8px; }
-    .controls button {
-      padding: 6px 14px;
-      font-family: var(--font-mono);
-      background: var(--color-bg-surface);
-      color: var(--color-fg-primary);
-      border: 1px solid var(--color-border-default);
-      border-radius: var(--r-1);
-      cursor: pointer;
-    }
-    .controls button:hover { border-color: var(--color-accent-fg); }
-    .task-list { list-style: none; padding: 0; margin: 0; }
-    .task {
-      display: grid;
-      grid-template-columns: 80px 1fr 80px 40px 90px;
-      gap: 8px;
-      padding: 6px 10px;
-      font-family: var(--font-mono);
-      font-size: var(--fs-12);
-      border-bottom: 1px solid var(--color-border-default);
-    }
-    .task--running { background: color-mix(in oklab, var(--ok), transparent 90%); }
-    .task--completed { opacity: 0.55; }
-    .task__id { color: var(--color-fg-muted); }
-    .task__priority { text-align: right; color: var(--color-accent-fg); }
-    footer code {
-      margin-right: 12px;
-      color: var(--color-fg-muted);
-    }
-  </style>
-</head>
-<body>
-  <div id="app"></div>
-  <script type="module" src="./solid-poc.tsx"></script>
-</body>
-</html>
