+++ dashboard/package.json
@@ -14,7 +14,7 @@
     "lint": "eslint src",
     "typecheck": "tsc --noEmit",
     "preview": "vite preview",
-    "test": "vitest run --config vitest.config.ts --no-file-parallelism --maxWorkers=1 && vitest run --config vitest.solid.config.ts --no-file-parallelism --maxWorkers=1",
+    "test": "vitest run --config vitest.config.ts --no-file-parallelism --maxWorkers=1",
     "test:serial": "pnpm test",
     "test:watch": "vitest",
     "tokens:build": "tsx design-system/tokens/build.ts",
@@ -43,13 +43,11 @@
     "ninja-keys": "^1.2.2",
     "preact": "^10.29.1",
     "shiki": "^4.0.2",
-    "solid-js": "^1.9.12",
     "valibot": "^1.3.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.39.4",
     "@preact/preset-vite": "^2.10.5",
-    "@solidjs/testing-library": "^0.8.10",
     "@tailwindcss/vite": "^4.2.4",
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/preact": "^3.2.4",
@@ -69,7 +67,6 @@
     "typescript": "^6.0.3",
     "typescript-eslint": "^8.59.2",
     "vite": "^7.3.5",
-    "vite-plugin-solid": "^2.11.12",
     "vitest": "^4.1.5"
   },
   "pnpm": {
